backend-src/requirements-fresh.txt
en-core-web-lg
nltk
pip-chill
scikit-learn
spacy-lookups-data
tensorflow
